// src/8th/features/library/model/book-info.ts
export type BookDetailInfo = {
  bookId: string
  title: string
  author: string
  publisher: string
  genre: string
  theme: string
  preferenceAverage: number
  bookLevel: string
  rgPoint: number
  readCount: number
  pages: number
  wordCount: number
  surfaceImage: string
  keywords: string
  grade: string
  readerLevel: string
  synopsis: string
  bookCode: string
  studentId: string
  studyId: string
  studentHistoryId: string
  studyStatus: string
  getRgPoint: number
  scoreStep1: number
  scoreStep2: number
  scoreStep3: number
  scoreStep4: number
  scoreStep5: number
  scoreStep6: number
  average: number
  registDate: string
  endDate: string
  fullEasyCode: string
  recentlyEndDate: string
  recentlyFullEasyCode: string
  passCount: number
  totalCompleteCount: number
  className: string
  fullEasyYn: boolean
  recommendedAge: string
  seriesName: string
  oldBookCode: string
  openDate: string
  studyStartedYn: boolean
  classId: string
  sampleYn: boolean
  modifyedRgPoint: number
  addStudyCount: number
  todayStudyYn: boolean
  assignedYn: boolean
  firstFullEasyCode: string
  secondFullEasyCode: string
  animationPath: string
  bookMarkYn: boolean
  backgroundImage: string
  writingStatusCode: string
  revisionStatusCode: string
  deleteYn: boolean
  price: number
  priceCd: string
  gameLandRoundOpenYn: boolean
  readingCompletedEB: string
  speakPassYn: boolean
  highlightDataYn: boolean
  infoType: string
  studyMode: string
  secondRgPoint: number
  userRgPoint: number
  studyTypeFullEasyYn: boolean
  speakContentYn: boolean
  levelRoundId: string
  workSheetPath: string
  vocabularyPath: string
  reportPath: string
  expiredContentsYn: boolean
}